+++ starkinfra/issuingpurchase/__init__.py
@@ -1,3 +1,3 @@
 from . import log
 from .log.__log import Log
-from .__issuingpurchase import query, get, parse, response
+from .__issuingpurchase import query, get, update, parse, response