+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Added
+- Update IssuingPurchase
 
 ## [0.10.1] - 2023-11-13
 ### Fixed